pyproject.toml
[project]
name = "trinket"
version = "0.1.0"
description = "Reusable Python snippets."
authors = [
    { name = "Adam Wen", email = "31234578+tzuhanwen@users.noreply.github.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyyaml (>=6.0.3,<7.0.0)"]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
ruff = "^0.11.6"
black = "^25.1.0"
mypy = "^1.15.0"
types-pyyaml = "^6.0.12.20250915"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "N", "C90", "S"]
ignore = ["E731"]
fixable = ["ALL"]
preview = true

[tool.mypy]
files = ["relic", "relic_cli", "tests", "hooks"]
python_version = "3.12"
strict = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
